Added CrossedArrayApllyOperation class, which performs operations optimised for different layout types of matrices. Made some changes to code style on whole repo

+++ Include/Types/Matrix.hpp
@@ -20,7 +20,11 @@
 template<typename NumType>
 class Matrix1;
 
-template<typename NumType, size_t ThreadCap = 20, size_t (*Decider)(size_t) = LogarithmicThreads<ThreadCap>>
+template<
+        typename NumType,
+        size_t ThreadCap = 20,
+        size_t (*Decider)(size_t) = LogarithmicThreads<ThreadCap>
+        >
 inline Matrix1<NumType> GetOuterProduct(const Vector<NumType>& A, const Vector<NumType>& B, bool HorizontalReturn = false);
 
 template<typename NumT>
@@ -89,17 +93,63 @@ class Matrix1 : public Vector<NumType>
 	void MoveFromPointer(NumType *Src, size_t SrcSoL = 0);
 public:
     // Basic constructors:
-    Matrix1(size_t Rows, size_t Cols, bool ByRow = false, ResourceManager* MM = DefaultMM);
-    Matrix1(std::initializer_list<std::initializer_list<NumType>> Init, bool ByRow = false, ResourceManager* MM = DefaultMM);
-    Matrix1(size_t Rows, size_t Cols, const NumType* Init, bool ByRow = false, ResourceManager* MM = DefaultMM);
-	Matrix1(const Matrix1& Target) noexcept;
-	Matrix1(Matrix1&& Target) noexcept;
+    Matrix1(
+        size_t Rows,
+        size_t Cols,
+        bool ByRow = false,
+        ResourceManager* MM = DefaultMM
+    );
+
+    Matrix1(
+        std::initializer_list<std::initializer_list<NumType>> Init,
+        bool ByRow = false,
+        ResourceManager* MM = DefaultMM
+    );
+
+    Matrix1(
+        size_t Rows,
+        size_t Cols,
+        const NumType* Init,
+        bool ByRow = false,
+        ResourceManager* MM = DefaultMM
+    );
+
+	Matrix1(
+        const Matrix1& Target
+    ) noexcept;
+
+    Matrix1(
+        Matrix1&& Target
+    ) noexcept;
 
     // Actually delegates construction to one of the above ones
-    Matrix1(size_t Rows, size_t Cols, NumType InitVal, bool ByRow = false, ResourceManager* MM = DefaultMM);
-    Matrix1(size_t NNSize, bool ByRow = false, ResourceManager* MM = DefaultMM);
-    Matrix1(size_t NNSize, NumType InitVal, bool ByRow = false, ResourceManager* MM = DefaultMM);
-	Matrix1(size_t NNSize, const NumType* Init, bool ByRow = false, ResourceManager* MM = DefaultMM);
+    Matrix1(
+        size_t Rows,
+        size_t Cols,
+        NumType InitVal,
+        bool ByRow = false,
+        ResourceManager* MM = DefaultMM
+    );
+
+    Matrix1(
+        size_t NNSize,
+        bool ByRow = false,
+        ResourceManager* MM = DefaultMM
+    );
+
+    Matrix1(
+        size_t NNSize,
+        NumType InitVal,
+        bool ByRow = false,
+        ResourceManager* MM = DefaultMM
+    );
+
+	Matrix1(
+        size_t NNSize,
+        const NumType* Init,
+        bool ByRow = false,
+        ResourceManager* MM = DefaultMM
+    );
 
 	// Data accessing operators
 
@@ -180,7 +230,11 @@ public:
 private:
     friend MatrixSumMachine<NumType> GetMSM<>(const Matrix1<NumType>& MatA, const Matrix1<NumType>& MatB, Matrix1<NumType>& MatC);
 public:
-    template<bool HorizontalReturn = false, size_t ThreadCap = 8, size_t (*Decider)(size_t) = LogarithmicThreads<ThreadCap>>
+    template<
+            bool HorizontalReturn = false,
+            size_t ThreadCap = 8,
+            size_t (*Decider)(size_t) = LogarithmicThreads<ThreadCap>
+    >
 	friend Matrix1 operator+(const Matrix1& A, const Matrix1& B)
     {
         if (A.Rows != B.Rows || A.Cols != B.Cols) [[unlikely]]
@@ -196,7 +250,10 @@ public:
 private:
     friend GPMM<NumType> GetMultMachine<>(const Matrix1<NumType> &A, const Matrix1<NumType> &B, const Matrix1<NumType> &C);
 public:
-    template<size_t ThreadCap = 20, size_t (*Decider)(size_t) = LogarithmicThreads<ThreadCap>>
+    template<
+            size_t ThreadCap = 20,
+            size_t (*Decider)(size_t) = LogarithmicThreads<ThreadCap>
+            >
 	friend Matrix1 operator*(const Matrix1& A, const Matrix1& B){
         if (A.Cols != B.Rows) [[unlikely]]
             throw std::runtime_error("[ERROR] Not able to perform matrix multiplication, wrong matrix sizes\n");
@@ -212,7 +269,10 @@ private:
     friend VMM<NumType> GetVMM<>(const Matrix1<NumType>& Mat, const Vector<NumType>& Vect, Vector<NumType>& RetVect);
 
 public:
-    template<size_t ThreadCap = 20, size_t (*Decider)(size_t) = LogarithmicThreads<ThreadCap>>
+    template<
+            size_t ThreadCap = 20,
+            size_t (*Decider)(size_t) = LogarithmicThreads<ThreadCap>
+            >
     friend Vector<NumType> operator*(const Matrix1& A, const Vector<NumType>& B){
         if (B.GetIsHorizontal() || A.Cols != B.GetSize())[[unlikely]]
             throw std::runtime_error("[ERROR] Not able to perform Vect and Matrix multiplication due to wrong sizes or dimensions\n");
@@ -225,7 +285,10 @@ public:
         return RetVal;
     }
 
-    template<size_t ThreadCap = 20, size_t (*Decider)(size_t) = LogarithmicThreads<ThreadCap>>
+    template<
+            size_t ThreadCap = 20,
+            size_t (*Decider)(size_t) = LogarithmicThreads<ThreadCap>
+            >
     friend Vector<NumType> operator*(const Vector<NumType>& A, const Matrix1& B){
         if (!A.GetIsHorizontal() || B.Rows != A.GetSize()) [[unlikely]]
             throw std::runtime_error("[ERROR] Not able to perform Vect and Matrix multiplication due to wrong sizes or dimensions\n");
@@ -253,7 +316,11 @@ public:
 private:
     friend OuterProductMachine<NumType> GetOPM<>(const Vector<NumType>& A, const Vector<NumType>& B, Matrix1<NumType>& C);
 public:
-    template<typename NumT, size_t ThreadCap, size_t (*Decider)(size_t)>
+    template<
+            typename NumT,
+            size_t ThreadCap,
+            size_t (*Decider)(size_t)
+            >
     friend Matrix1<NumT> GetOuterProduct(const Vector<NumT>& A, const Vector<NumT>& B, bool HorizontalReturn);
 
     friend Matrix1<NumType> ElemByElemMult(const Matrix1<NumType>& A, const Matrix1<NumType>& B){
@@ -269,6 +336,7 @@ public:
         }
     }
 
+    // TODO: make directly apply elements but with matrix
 //    Vector<NumType>& ApplyElemByElemMult(const Vector<NumType>& B){
 //        if (Size != B.Size) [[unlikely]]{
 //            throw std::runtime_error("[ERROR] Not able to perform ApplyElemByElemMult, because Vectors have different sizes\n");
@@ -299,25 +367,65 @@ public:
 
 template<typename NumT>
 inline MatrixSumMachine<NumT> GetMSM(const Matrix1<NumT>& MatA, const Matrix1<NumT>& MatB, Matrix1<NumT>& MatC){
-    return MatrixSumMachine<NumT>(MatA.Array, MatB.Array, MatC.Array, MatA.Rows, MatA.Cols, MatA.Size, MatA.SizeOfLine,
-                                  MatB.SizeOfLine, MatC.SizeOfLine, MatA.IsHorizontal, MatB.IsHorizontal, MatC.IsHorizontal);
+    return MatrixSumMachine<NumT>(
+            MatA.Array,
+            MatB.Array,
+            MatC.Array,
+            MatA.Rows,
+            MatA.Cols,
+            MatA.Size,
+            MatA.SizeOfLine,
+            MatB.SizeOfLine,
+            MatC.SizeOfLine,
+            MatA.IsHorizontal,
+            MatB.IsHorizontal,
+            MatC.IsHorizontal
+    );
 }
 
 template<typename NumT>
 inline OuterProductMachine<NumT> GetOPM(const Vector<NumT>& A, const Vector<NumT>& B, Matrix1<NumT>& C)
 {
-    return OuterProductMachine<NumT>(A.GetArray(), B.GetArray(), C.Array, A.GetSize(), B.GetSize(), C.SizeOfLine, C.GetIsHorizontal());
+    return OuterProductMachine<NumT>(
+            A.GetArray(),
+            B.GetArray(),
+            C.Array,
+            A.GetSize(),
+            B.GetSize(),
+            C.SizeOfLine,
+            C.GetIsHorizontal()
+    );
 }
 
 template<typename NumT>
 VMM<NumT> GetVMM(const Matrix1<NumT> &Mat, const Vector<NumT> &Vect, Vector<NumT> &RetVect) {
-    return VMM<NumT>(Mat.Array, Vect.GetArray(), RetVect.GetArray(), Mat.Rows, Mat.Cols, Mat.SizeOfLine, Mat.IsHorizontal);
+    return VMM<NumT>(
+            Mat.Array,
+            Vect.GetArray(),
+            RetVect.GetArray(),
+            Mat.Rows,
+            Mat.Cols,
+            Mat.SizeOfLine,
+            Mat.IsHorizontal
+    );
 }
 
 template<typename NumT>
 GPMM<NumT> GetMultMachine(const Matrix1<NumT> &A, const Matrix1<NumT> &B, const Matrix1<NumT> &C) {
-    return GPMM<NumT>(A.Array, B.Array, C.Array, A.Rows, A.Cols, B.Cols, A.SizeOfLine, B.SizeOfLine, C.SizeOfLine,
-                      A.IsHorizontal, B.IsHorizontal, C.IsHorizontal);
+    return GPMM<NumT>(
+            A.Array,
+            B.Array,
+            C.Array,
+            A.Rows,
+            A.Cols,
+            B.Cols,
+            A.SizeOfLine,
+            B.SizeOfLine,
+            C.SizeOfLine,
+            A.IsHorizontal,
+            B.IsHorizontal,
+            C.IsHorizontal
+    );
 }
 
 // ------------------------------------
@@ -378,9 +486,16 @@ void Matrix1<NumType>::SetupAccess()
 }
 
 template<typename NumType>
-Matrix1<NumType>::Matrix1(size_t Rows, size_t Cols, const NumType *Init, bool ByRow, ResourceManager *MM) :
-        Vector<NumType>{ ByRow, MM }, Rows{ Rows }, Cols{ Cols },
-        MatrixSize{ (unsigned long)Rows * (unsigned long)Cols }
+Matrix1<NumType>::Matrix1(
+        size_t Rows,
+        size_t Cols,
+        const NumType *Init,
+        bool ByRow, ResourceManager *MM
+):
+        Vector<NumType>{ ByRow, MM },
+        Rows{ Rows },
+        Cols{ Cols },
+        MatrixSize{ Rows * Cols }
 {
     PerformSanityChecks();
     SetupAccess();
@@ -389,12 +504,26 @@ Matrix1<NumType>::Matrix1(size_t Rows, size_t Cols, const NumType *Init, bool By
 }
 
 template<typename NumType>
-Matrix1<NumType>::Matrix1(size_t NNSize, const NumType *Init, bool ByRow, ResourceManager *MM) :
-    Matrix1(NNSize, NNSize, Init, ByRow, MM) {}
+Matrix1<NumType>::Matrix1(
+        size_t NNSize,
+        const NumType *Init,
+        bool ByRow,
+        ResourceManager *MM
+):
+    Matrix1(
+            NNSize,
+            NNSize,
+            Init,
+            ByRow,
+            MM
+    ) {}
 
 template<typename NumType>
-Matrix1<NumType>::Matrix1(std::initializer_list<std::initializer_list<NumType>> Init, bool ByRow,
-                          ResourceManager *MM) :
+Matrix1<NumType>::Matrix1(
+        std::initializer_list<std::initializer_list<NumType>> Init,
+        bool ByRow,
+        ResourceManager *MM
+):
         Vector<NumType>(ByRow, MM)
 {
     const std::initializer_list<NumType>* InitData = std::data(Init);
@@ -425,37 +554,79 @@ Matrix1<NumType>::Matrix1(std::initializer_list<std::initializer_list<NumType>> 
 }
 
 template<typename NumType>
-Matrix1<NumType>::Matrix1(Matrix1 &&Target) noexcept :
-        Vector<NumType>{std::move((Vector<NumType>&&)Target) }, Rows{ Target.Rows }, Cols{ Target.Cols } ,  MatrixSize{Target.MatrixSize },
-        OffsetPerLine{ Target.OffsetPerLine },  SizeOfLine{ Target.SizeOfLine }, IsMemoryPacked{ Target.IsMemoryPacked }
+Matrix1<NumType>::Matrix1(
+        Matrix1 &&Target
+) noexcept:
+        Vector<NumType>{std::move((Vector<NumType>&&)Target) },
+        Rows{ Target.Rows },
+        Cols{ Target.Cols } ,
+        MatrixSize{Target.MatrixSize },
+        OffsetPerLine{ Target.OffsetPerLine },
+        SizeOfLine{ Target.SizeOfLine },
+        IsMemoryPacked{ Target.IsMemoryPacked }
 {
     SetupAccess();
 }
 
 template<typename NumType>
-Matrix1<NumType>::Matrix1(const Matrix1 &Target) noexcept :
-        Vector<NumType>(Target), Rows{ Target.Rows }, Cols{ Target.Cols },
-        MatrixSize{ Target.MatrixSize }, IsMemoryPacked{ Target.IsMemoryPacked },
-        OffsetPerLine{ Target.OffsetPerLine }, SizeOfLine{ Target.SizeOfLine }
+Matrix1<NumType>::Matrix1(
+        const Matrix1 &Target
+) noexcept:
+        Vector<NumType>(Target),
+        Rows{ Target.Rows },
+        Cols{ Target.Cols },
+        MatrixSize{ Target.MatrixSize },
+        IsMemoryPacked{ Target.IsMemoryPacked },
+        OffsetPerLine{ Target.OffsetPerLine },
+        SizeOfLine{ Target.SizeOfLine }
 {
     SetupAccess();
 }
 
 template<typename NumType>
-Matrix1<NumType>::Matrix1(size_t Rows, size_t Cols, NumType InitVal, bool ByRow, ResourceManager *MM)  :
-          Matrix1(Rows, Cols, ByRow, MM)
+Matrix1<NumType>::Matrix1(
+        size_t Rows,
+        size_t Cols,
+        NumType InitVal,
+        bool ByRow,
+        ResourceManager *MM
+):
+          Matrix1(
+              Rows,
+              Cols,
+              ByRow,
+              MM
+          )
 {
     Vector<NumType>::SetWholeData(InitVal);
 }
 
 template<typename NumType>
-Matrix1<NumType>::Matrix1(size_t NNSize, NumType InitVal, bool ByRow, ResourceManager *MM):
-    Matrix1(NNSize, NNSize, InitVal, ByRow, MM) {}
+Matrix1<NumType>::Matrix1(
+        size_t NNSize,
+        NumType InitVal,
+        bool ByRow,
+        ResourceManager *MM
+):
+    Matrix1(
+            NNSize,
+            NNSize,
+            InitVal,
+            ByRow,
+            MM
+    ) {}
 
 template<typename NumType>
-Matrix1<NumType>::Matrix1(size_t Rows, size_t Cols, bool ByRow, ResourceManager *MM)  :
-        Vector<NumType>{ ByRow, MM }, Rows{ Rows }, Cols{ Cols },
-        MatrixSize { (unsigned long)Rows * (unsigned long)Cols }
+Matrix1<NumType>::Matrix1(
+        size_t Rows,
+        size_t Cols,
+        bool ByRow,
+        ResourceManager *MM
+):
+        Vector<NumType>{ ByRow, MM },
+        Rows{ Rows },
+        Cols{ Cols },
+        MatrixSize { Rows * Cols }
 
 {
     PerformSanityChecks();
@@ -464,8 +635,17 @@ Matrix1<NumType>::Matrix1(size_t Rows, size_t Cols, bool ByRow, ResourceManager 
 }
 
 template<typename NumType>
-Matrix1<NumType>::Matrix1(size_t NNSize, bool ByRow, ResourceManager *MM)  :
-    Matrix1(NNSize, NNSize, ByRow, MM){}
+Matrix1<NumType>::Matrix1(
+        size_t NNSize,
+        bool ByRow,
+        ResourceManager *MM
+):
+    Matrix1(
+            NNSize,
+            NNSize,
+            ByRow,
+            MM
+    ){}
 
 template<typename NumType>
 void Matrix1<NumType>::MoveFromPointer(NumType * const Src, size_t SrcSoL)
